+++ great_expectations/metrics/__init__.py
@@ -1,15 +1,15 @@
 from .batch.batch_column_types import BatchColumnTypes
 from .batch.row_count import BatchRowCount
-from .column.column_descriptive_stats import ColumnDescriptiveStats
-from .column.column_distinct_values import ColumnDistinctValues
-from .column.column_distinct_values_count import ColumnDistinctValuesCount
-from .column.column_null_count import ColumnNullCount
-from .column.column_sample_values import ColumnSampleValues
-from .column.column_values_match_regex_count import ColumnValuesMatchRegexCount
-from .column.column_values_match_regex_values import ColumnValuesMatchRegexValues
+from .column.descriptive_stats import ColumnDescriptiveStats
+from .column.distinct_values import ColumnDistinctValues
+from .column.distinct_values_count import ColumnDistinctValuesCount
 from .column.mean import ColumnMean
+from .column.null_count import ColumnNullCount
+from .column.sample_values import ColumnSampleValues
+from .column.values_match_regex_count import ColumnValuesMatchRegexCount
+from .column.values_match_regex_values import ColumnValuesMatchRegexValues
 from .column.values_non_null import ColumnValuesNonNull, ColumnValuesNonNullCount
-from .column_pair.values_in_set import ColumnPairValuesInSet
+from .column_pair.values_in_set import ColumnPairValuesInSetUnexpectedCount
 from .metric import Metric
 from .multi_column.sum_equal import MultiColumnSumEqualUnexpectedCount
 from .query.row_count import QueryRowCount
