+++ tests/integration/metrics/column/test_distinct_values.py
@@ -1,7 +1,7 @@
 import pandas as pd
 
 from great_expectations.datasource.fluent.interfaces import Batch
-from great_expectations.metrics.column.column_distinct_values import (
+from great_expectations.metrics.column.distinct_values import (
     ColumnDistinctValues,
     ColumnDistinctValuesResult,
 )
