+++ tests/integration/metrics/column/test_sample_values.py
@@ -1,7 +1,7 @@
 import pandas as pd
 
 from great_expectations.datasource.fluent.interfaces import Batch
-from great_expectations.metrics.column.column_sample_values import (
+from great_expectations.metrics.column.sample_values import (
     ColumnSampleValues,
     ColumnSampleValuesResult,
 )
