+++ tests/integration/metrics/column/test_values_match_regex_count.py
@@ -1,7 +1,7 @@
 import pandas as pd
 
 from great_expectations.datasource.fluent.interfaces import Batch
-from great_expectations.metrics.column.column_values_match_regex_count import (
+from great_expectations.metrics.column.values_match_regex_count import (
     ColumnValuesMatchRegexCount,
     ColumnValuesMatchRegexCountResult,
 )
